refactor: renombra archivos para utilizar singular
usersController por userController
usersModel por userModel
usersRoute por userRoute
